Updated database and unnecessary code

+++ database/migrations/2019_09_28_185945_create_activity_event_table.php
@@ -16,10 +16,10 @@ class CreateActivityEventTable extends Migration
         Schema::create('activity_event', function (Blueprint $table) {
             $table->bigIncrements('id');
 
-            $table->unsignedBigInteger('activity_id');
+            $table->unsignedInteger('activity_id');
             $table->foreign('activity_id')->references('id')->on('activities')->onDelete('cascade');
 
-            $table->unsignedBigInteger('event_id');
+            $table->unsignedInteger('event_id');
             $table->foreign('event_id')->references('id')->on('events')->onDelete('cascade');
 
             $table->timestamps();
